add to cart implemented

+++ src/app/my-account/profile/profile.component.ts
@@ -13,7 +13,12 @@ export class ProfileComponent implements OnInit {
   patientForm: FormGroup
   submitted: boolean = false
   patients: any = []
-  bloodGroup: any  =[]
+  bloodGroup: any = []
+  patientId:any
+  isEdit: boolean = false
+  username:any = ''
+
+
 
   constructor(private _fb: FormBuilder,
     private _profile: ProfileService) {
@@ -31,6 +36,7 @@ export class ProfileComponent implements OnInit {
   }
 
   ngOnInit(): void {
+    this.username = localStorage.getItem('USER_NAME')
     let payload = {
       schemaName: 'nir1691144565',
       user_id: localStorage.getItem('USER_ID')
@@ -41,8 +47,8 @@ export class ProfileComponent implements OnInit {
       }
     })
 
-    this._profile.getBloodGroup().subscribe((res:any) => {
-      if(res.status == 1) {
+    this._profile.getBloodGroup().subscribe((res: any) => {
+      if (res.status == 1) {
         this.bloodGroup = res.data
       }
     })
@@ -68,7 +74,7 @@ export class ProfileComponent implements OnInit {
       this._profile.storePatient(payload).subscribe((res: any) => {
         if (res.status == 1) {
           Swal.fire({
-            position: 'top-end',
+            position: 'center',
             icon: 'success',
             text: 'Added Successfully!',
             showConfirmButton: false,
@@ -76,7 +82,7 @@ export class ProfileComponent implements OnInit {
           })
           $("#exampleModal").hide();
           $('body').removeClass('modal-open');
-          // $(".modal-backdrop").removeClass("modal-backdrop show");
+          $(".modal-backdrop").removeClass("modal-backdrop show");
           this.ngOnInit()
         } else {
           Swal.fire({
@@ -89,15 +95,87 @@ export class ProfileComponent implements OnInit {
     }
   }
 
-  deleteItem(id:any) {
+  editPatient(id: any) {
+    this.isEdit = true
+    this.patientId = id
     let payload = {
       "schemaName": "nir1691144565",
       "patient_id": id
-  }
-  this._profile.deletePatient(payload).subscribe((res:any) => {
-    if(res.status == 1) {
-      this.ngOnInit()
     }
-  })
+    this._profile.getPatientById(payload).subscribe((res: any) => {
+      if (res.status == 1) {
+        console.log(res)
+        this.patientForm.get("patientName").setValue(res.data.patient_name);
+        this.patientForm.get("age").setValue(res.data.age);
+        this.patientForm.get("blood_group").setValue(res.data.blood_group);
+        this.patientForm.get("gender").setValue(res.data.gender);
+        this.patientForm.get("dob").setValue(res.data.dob);
+        this.patientForm.get("height").setValue(res.data.height);
+        this.patientForm.get("weight").setValue(res.data.weight);
+      }
+    })
+  }
+
+  updatePatient() {
+    this.submitted = true
+    let form = this.patientForm.value
+    let payload = {
+      "schemaName": "nir1691144565",
+      "patient_id": Number(this.patientId),
+      "patientName": form.patientName,
+      "age": Number(form.age),
+      "blood_group": 1,
+      "gender": Number(form.gender),
+      "dob": form.dob,
+      "height": Number(form.height),
+      "weight": Number(form.weight)
+    }
+    this._profile.updatePatients(payload).subscribe((res:any) => {
+      if(res.status == 1) {
+        Swal.fire({
+          position: 'center',
+          icon: 'success',
+          text: 'Added Successfully!',
+          showConfirmButton: false,
+          timer: 1500
+        })
+        $("#exampleModal").hide();
+        $('body').removeClass('modal-open');
+        this.ngOnInit();
+        this.isEdit = false
+      } else {
+        Swal.fire({
+          icon: 'error',
+          title: 'Oops...',
+          text: 'Something went wrong!',
+        })
+      }
+    })
+  }
+
+  deleteItem(id: any) {
+    let payload = {
+      "schemaName": "nir1691144565",
+      "patient_id": id
+    }
+
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You want to remove this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Remove'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this._profile.deletePatient(payload).subscribe((res: any) => {
+          if (res.status == 1) {
+            this.ngOnInit()
+          }
+        })
+      }
+    })
+
   }
 }
